build: refactor to class
