+++ src/cli.ts
@@ -1,11 +1,9 @@
-import {
-  generateStateProof,
-  generateTransactionProof,
-  generateTxReceiptProof,
-  setupEthRPCClient,
-} from './generate-proof'
+import { ProofGenerator } from './proof-generator'
 
 const args = process.argv.slice(2)
+const proofGenerator = new ProofGenerator(
+  'https://sepolia.infura.io/v3/ee1e6d7e77c2415386766fa559769941',
+)
 
 if (args.length) {
   switch (args[0]) {
@@ -16,7 +14,8 @@ if (args.length) {
         )
         process.exit(1)
       }
-      generateStateProof(args[1], args[2], args[3], setupEthRPCClient())
+      proofGenerator
+        .generateStateProof(args[1], args[2], args[3])
         .then(console.log)
         .catch(console.error)
       break
@@ -26,7 +25,8 @@ if (args.length) {
         console.error('Usage: node script.js receipt <txId>')
         process.exit(1)
       }
-      generateTxReceiptProof(args[1], setupEthRPCClient())
+      proofGenerator
+        .generateTxReceiptProof(args[1])
         .then(console.log)
         .catch(console.error)
       break
@@ -36,7 +36,8 @@ if (args.length) {
         console.error('Usage: node script.js transaction <txId>')
         process.exit(1)
       }
-      generateTransactionProof(args[1], setupEthRPCClient())
+      proofGenerator
+        .generateTransactionProof(args[1])
         .then(console.log)
         .catch(console.error)
       break
